llm/requirements.txt
aiohttp[speedups]
langchain-google-genai
python-dotenv